+++ email/src/watch/config.rs
@@ -1,6 +1,7 @@
+use futures::Future;
 use process::Cmd;
 use serde::{Deserialize, Serialize};
-use std::{fmt, ops::Deref, sync::Arc};
+use std::{fmt, ops::Deref, pin::Pin, sync::Arc};
 
 use crate::{envelope::Envelope, Result};
 
@@ -8,18 +9,18 @@ use crate::{envelope::Envelope, Result};
 ///
 /// Each variant represent the action that should be done when a
 /// change occurs.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "kebab-case")]
-pub enum WatchHook {
+pub struct WatchHook {
     /// Execute the shell command.
     ///
     /// For now, command is executed without any parameter nor
     /// input. This may change in the future.
-    Cmd(Cmd),
+    pub cmd: Option<Cmd>,
 
     /// Send a system notification using the given
     /// [`notify_rust::Notification`]-like configuration.
-    Notify(WatchNotifyConfig),
+    pub notify: Option<WatchNotifyConfig>,
 
     /// Execute the given watch function.
     ///
@@ -27,7 +28,17 @@ pub enum WatchHook {
     /// should take a reference to an envelope and return a [`Result`]
     /// of unit.
     #[serde(skip)]
-    Fn(WatchFn),
+    pub callback: Option<WatchFn>,
+}
+
+impl Eq for WatchHook {
+    //
+}
+
+impl PartialEq for WatchHook {
+    fn eq(&self, other: &Self) -> bool {
+        self.cmd == other.cmd && self.notify == other.notify
+    }
 }
 
 /// Watch function.
@@ -35,17 +46,28 @@ pub enum WatchHook {
 /// This is just a wrapper around a function that takes a reference to
 /// an envelope.
 #[derive(Clone)]
-pub struct WatchFn(Arc<dyn Fn(&Envelope) -> Result<()> + Send + Sync>);
+pub struct WatchFn(
+    Arc<dyn Fn(&Envelope) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync>,
+);
 
 impl WatchFn {
     /// Create a new watch function.
-    pub fn new(f: impl Fn(&Envelope) -> Result<()> + Send + Sync + 'static) -> Self {
-        Self(Arc::new(f))
+    pub fn new<F: Future<Output = Result<()>> + Send + 'static>(
+        f: impl Fn(&Envelope) -> F + Send + Sync + 'static,
+    ) -> Self {
+        Self(Arc::new(move |envelope| Box::pin(f(envelope))))
+    }
+}
+
+impl Default for WatchFn {
+    fn default() -> Self {
+        Self(Arc::new(|_| Box::pin(async { Ok(()) })))
     }
 }
 
 impl Deref for WatchFn {
-    type Target = Arc<dyn Fn(&Envelope) -> Result<()> + Send + Sync>;
+    type Target =
+        Arc<dyn Fn(&Envelope) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -54,17 +76,7 @@ impl Deref for WatchFn {
 
 impl fmt::Debug for WatchFn {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "WatchHookFn()")
-    }
-}
-
-impl Eq for WatchFn {
-    //
-}
-
-impl PartialEq for WatchFn {
-    fn eq(&self, _other: &Self) -> bool {
-        true
+        write!(f, "WatchFn()")
     }
 }